Removed methods to load in data from tasks.py
